gpu_metrics: update get_average_socket_power type to Option<u32>

+++ amdgpu/gpu_metrics.rs
@@ -89,7 +89,7 @@ impl MetricsInfo for GpuMetrics {
     impl_metrics!(get_average_umc_activity, Option<u16>);
     impl_metrics!(get_average_mm_activity, Option<u16>);
     impl_metrics!(get_system_clock_counter, Option<u64>);
-    impl_metrics!(get_average_socket_power, Option<u16>);
+    impl_metrics!(get_average_socket_power, Option<u32>);
     impl_metrics!(get_average_cpu_power, Option<u16>);
     impl_metrics!(get_average_soc_power, Option<u16>);
     impl_metrics!(get_average_gfx_power, Option<u16>);
